new speed option
Key changes made:

Added a new "Manual" option to the speed dropdown in the HTML.
Introduced a manualFlipMode variable to track when the manual flip option is selected.
Modified the flipCard function to handle the manual flip mode:
If in manual mode and cards are already flipped, it resets them before flipping the new card.
Updated the checkMatch function to not automatically reset unmatched cards when in manual mode.
Modified the resetFlip function to only reset unmatched cards.
Updated the loadGameFromURL function to handle the new manual flip mode option.
These changes allow for a fifth option where the tiles don't automatically flip back, but instead wait for the next card to be clicked before resetting. This gives players more time to memorize the cards if they choose this option.